Implement strict and lenient modes for Kotlin metadata reading
In order to handle potential cases of inconsistencies in Kotlin metadata, we've introduced strict and lenient modes for reading Kotlin metadata. This change provides much-needed flexibility in terms of how strict or lenient the handling of Kotlin metadata can be. In strict mode, an exception will be thrown when inconsistent metadata is encountered. In lenient mode, the reader will attempt to handle the inconsistent metadata by ignoring certain inconsistencies.

Update Kotlin metadata reading in kapt3 compiler

Changed the "read()" function to "readStrict()" in the line rendering the metadata for Kotlin classes in the kapt3 compiler. This is part of implementing a more strict and lenient control for reading Kotlin metadata. The strict mode, used here, will throw an exception when inconsistencies are encountered, thus providing robust error detection.