FIR entrypoint: deduplicate some code and remove JVM dependency
Deduplicate Fir2IrExtensions creation and move the
convertToIrAndActualizeForJvm function to 'cli', which allows to remove
the dependency of 'fir.entrypoint' on 'backend.jvm'.

Note that behavior slightly changes in FirKotlinToJvmBytecodeCompiler:
previously the value of linkViaSignatures was always false, now it is
taken from the compiler configuration, which seems more correct.